Changed AJAX code
All AJAX codes are using jquery post/get instead of the custom one now.
One problem which I didn't know if existed before is that the chat,
after entering, does not show immediately and instead user needs to wait
for the refresh delay to refresh it to see the new message. Also removed
caching codes which doesn't seem to make sense to me.